moved classes + interfaces into separate files
