+++ Assets/GlitchEscape/Scripts/Effects/BasicEffectBehavior.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 462eb1c360cc4eb289ac024481bc1040
+timeCreated: 1587323956
