+++ Assets/GlitchEscape/Scripts/Effects/IEffect.cs
@@ -1,0 +1,8 @@
+namespace GlitchEscape.Effects {
+    public interface IEffect {
+        IEffectBehavior effectBehavior { get; set; }
+        bool active { get; set; }
+        bool finished { get; }
+        void Cancel();
+    }
+}
