+++ Assets/GlitchEscape/Scripts/Effects/IEffect.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 22704ecdc4c247cdb2766fee8d0cf7de
+timeCreated: 1587323793
