+++ Assets/GlitchEscape/Scripts/Effects/IEffectBehavior.cs
@@ -1,0 +1,8 @@
+namespace GlitchEscape.Effects {
+    public interface IEffectBehavior {
+        bool active { get; set; }
+        bool finished { get; }
+        void OnCancelled();
+        void Update();
+    }
+}
