+++ Assets/GlitchEscape/Scripts/Effects/IEffectBehavior.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 678d5f9e789540b6adcf0f041497fc5f
+timeCreated: 1587323811
