+++ Assets/GlitchEscape/Scripts/Effects/IEffector.cs
@@ -1,0 +1,5 @@
+namespace GlitchEscape.Effects {
+    public interface IEffector<TOwner, TState> where TState : EffectState<TOwner, TState> {
+        void Apply(TState state);
+    }
+}
