+++ Assets/GlitchEscape/Scripts/Effects/IEffector.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f9161fc5d2944b52a81fe8cbf4ec6550
+timeCreated: 1587323903
